moved tour into own component, checked classnames and tesed locally, looks like it works
